requirements.txt
# FastAPI
fastapi
uvicorn[standard]
python-multipart

# Supabase
supabase

# Environment variables
python-dotenv

# Data validation
pydantic-settings

# Development (optional)
httpx
requests
